Add deck and list decks.
Create AddDeck component that adds a new deck to the local
component state.

List out existing decks, and newly created ones with
the number of questions (cards) available for each deck.

I am using local component state instead of redux for this
app. I will be saving the data to AsyncStorage though in
the future.