Setup AsyncStorage and Redux.
Initial setup of AsyncStorage to save card titles
to the DB.

Wireup redux instead of only trying to use local state
and prop hoisting. This became troublesome when I tried
to setup the StackNavigator and pass props down to child
components. It seemed much easier to setup redux for a
central store and the ability to use connect to access
the store in child components.