Allow deleting of decks.
Make sure decks are primed with empty
object if there are no decks in the action.

In the reducer only return the newly recieved
decks and ignore existing state.

Guard against trying to display a deck that
has been deleted. This was causing an error
as it would try to display the deleted
deck title then redirect quickly."